Extracted recording specific code out of src/emu/video.cpp and put into src/emu/recording.cpp
